+++ test/external/fuzz_uops.py
@@ -3,7 +3,7 @@ from collections import defaultdict
 import numpy as np
 from dataclasses import replace
 from typing import DefaultDict, Dict, List, Tuple
-from tinygrad.ops import END_FOR_UOP, UOp
+from tinygrad.codegen.uops import END_FOR_UOP, UOp
 from tinygrad.codegen.uopgraph import UOpGraph
 from tinygrad.device import Buffer, Device
 from tinygrad.engine.realize import CompiledRunner
