+++ test/test_dtype_alu.py
@@ -4,11 +4,12 @@ from tinygrad import Tensor, dtypes, Device
 import operator
 import numpy as np
 from hypothesis import given, strategies as strat, settings
+from tinygrad.codegen.uops import UOps
 from tinygrad.dtype import DType
 from tinygrad.helpers import CI, getenv
 from tinygrad.engine.schedule import create_schedule
 from tinygrad.engine.realize import run_schedule
-from tinygrad.ops import UnaryOps, UOps
+from tinygrad.ops import UnaryOps
 from tinygrad.tensor import _to_np_dtype
 from test.helpers import is_dtype_supported
