+++ tinygrad/codegen/lowerer.py
@@ -5,7 +5,8 @@ import functools
 from tinygrad.shape.shapetracker import ShapeTracker, View
 from tinygrad.shape.symbolic import sint
 from tinygrad.dtype import dtypes, PtrDType, ImageDType, DType
-from tinygrad.ops import ReduceOps, KernelInfo, BinaryOps, BUFFER_UOPS, UOp, UOps
+from tinygrad.ops import ReduceOps, KernelInfo, BinaryOps
+from tinygrad.codegen.uops import BUFFER_UOPS, UOp, UOps
 from tinygrad.renderer import Renderer
 from tinygrad.helpers import getenv, all_int, get_contraction, prod, partition, flatten
