+++ tinygrad/engine/realize.py
@@ -3,7 +3,8 @@ import time, pprint
 from collections import defaultdict
 from dataclasses import dataclass, replace
 from tinygrad.helpers import colored, getenv, DEBUG, GlobalCounters, ansilen, BEAM, NOOPT, all_int, CAPTURING, Metadata, Context, TRACEMETA, dedup
-from tinygrad.ops import MetaOps, UOps, UOp
+from tinygrad.codegen.uops import UOps, UOp
+from tinygrad.ops import MetaOps
 from tinygrad.dtype import dtypes
 from tinygrad.device import Device, Buffer
 from tinygrad.shape.symbolic import Variable, sym_infer, sint
