+++ tinygrad/renderer/__init__.py
@@ -2,7 +2,8 @@ from typing import Optional, List, Tuple, Dict, Callable, Any
 import functools
 from dataclasses import dataclass, field
 from tinygrad.helpers import to_function_name, dedup
-from tinygrad.ops import Op, UOps, UOp, flops_mem
+from tinygrad.codegen.uops import UOps, UOp, flops_mem
+from tinygrad.ops import Op
 from tinygrad.shape.symbolic import sym_infer, sint, Variable
 from tinygrad.dtype import DType
